DataBase Dao and Entity User added to the build
New Package "database" created.
The Package contains:
- ArchelonDb database class
-  ArchelonDbDao DAO class
- User @entity data class

The classes at the moment offer just the functionalities needed for User Login.

ArchelonDb after creating the Database, executes the method populateDatabase to populate the database with a user entity on the first run of the app.

ArchelonDbDao provide methods to insert update and fetch a user entity from ArchelonDb

The data fileds of the User Entity class are:
- email
- password
- userId