Spinners Introduced in the UI Elements - Wind Class Added to the DB Schema
Spinners replace EditTexts to help the user selection for:
- Beach
- Beach Sector
- Wind
- Precipitation
- Sky
In the Morning Survey UI Sequence.

Fixed the problem in the manifest file that caused app crashes on start-up after ArchelonApplication file is been removed.

 getters methods added to the dao class to pull the data from the db, getters methods added to the repository.
getters methods added to the ViewModel.

most of the entities objects override the toString method to be presented in the UI