Decoupled Login
The Login process is now handled by observing a MutableLivedata of the ViewModel, offering more decoupling and separation of concerns. The View doesn't make direct methods calls to
retrieve the login data anymore.
The View calls a login method of the ViewModel Instance.
The implementation of the Login method in the ViewModel has not been finished yet.

My Apologies about the messy comments and and commented line of code but work in progress.