Observers Entity table split in two and added to the Database Schema. More Spinners in Morning Survey for Morning Survey Leader and Observer
Observer Entity table splitted in two tables:
- Leaders for the morning Survey
  Leaders
- Observers for the 2nd Observers.

Dao and repository methods for insert in and pul out data from the db tables added.

Added spinners for helping the user selection of MS Leader and Observer in the fragments.

Cancel function added to the ViewModel to manage the case where  the user quit the insert of a new morning survey pressing the cancel button in the UI Sequence.